Removing connection pooling because there is no heartbeat mechanism available.
The HiveServer disconnects the connection if it does
not receive any query within five minutes approx. Currently
there is no mechanism in thrift library to send the heartbeat.
So connection pooling not an option for now and is removed.
Connection pooling will be back when we find a way to send
heartbeat or be able to detect disconnection.

Reported by Ahsan Hadi, Fix by me.

Issue (#3)