Fix for issue #54 and #55. This patch fixes two issues: a) A server crash when date and time data types are used b) Wrong results are returned on certain queries in which    parametrized where clause is pushed down.
For (a) The problem was in the handling of date and time
data types while passing them from C to Java.
The correct method is to send them as string and then use
valueOf Java method to get the actual data type value.

For (b) the problem was in the parameter number specified
while binding the parameter value.
The parameter number was not being passed by the caller
and the Java function was using its own class member to
generate the parameter number. This logic was error prone
and it was not working right either.
The correct method is to let the caller pass the parameter
number to bind, because the caller always knows the
correct parameter number to bind.

This commit fixes the following RMs
42813, 42782, 42774, 42771