Fix for issue #56, Error in queries needing more than one parameter to bind
The problem was with the logic in deparseParam function.
The function was assuming that the remote hive server would
behave like a remote Postgres server, where if a parameter
is used more than once in where clause it is counted as one
and the remote query uses $1 as many times as it is required to be
used in the query. This is an optimization where if a query needs
to use a parameter more than once in the where clause it is bound
once and used as many times as it is required.
This is not true for hive which uses ? as a parameter marker.
Clearly this optimization cannot be used with such a notation
and here we have to bind a parameter as many times as it is being
used in the query. The function deparseParam is therefore
rewritten keeping this fact into consideration.

Example:
In Postgres FDW the following query
explain verbose SELECT empno, ename, sal, deptno
        FROM emp e WHERE hiredate IN (
        SELECT hiredate FROM emp WHERE  emp.deptno < e.deptno OR
          emp.deptno > e.deptno);
generates the following remote query:
SELECT hiredate FROM public.emp WHERE
   (((deptno < $1::integer) OR (deptno > $1::integer)))

Note that it uses $1 twice in the query, meaning that only one
parameter is to be bound but used twice in the query.
This is fine for a remote postgres server, but consider the case
of a remote hive server where the same query generates the following
remote query:
SELECT hiredate FROM default.emp WHERE (((deptno < ?) OR (deptno > ?)))
Clearly here the "bind once use many" optimization cannot be used.