Fix for RM # 42861, provide a function to return code version
Add a SQL function to return code version
The function usage is as follows:

	select hdfs_fdw_version();
	 hdfs_fdw_version
	------------------
    	        20003
	(1 row)

The purpose of code version is as follows:
If the .so file is updated and the customer
just installs the update but does not do an
ALTER EXTENSION then the \dx version will still
be older but the .so will be newer.
To detect which version of .so is being used
we need a code version.