Fix tablename, column name to have quoted identifiers.
To access a table or column name having special characters or a
reserved word, it needs to be sent to remote server as a quoted
identifier. Hive sql expected a '`' character unlike '"' in case
postgresql. This commit adds code to handle this hive sql quoting.

FDW-293, Jeevan Ladhe, reviewed by Vaibhav Dalvi.