Push down joins to the remote HDFS servers.
If we have a join between two foreign tables from the same remote
server, push that join down to the remote server instead of fetching
all the rows for both the tables and performing a join locally.  This
will have a very significant win in many practical cases.  Whole-row
references in the join relations are handled too.

Since it is very unclear that which of the operators are supported on
the Hive/Spark servers, currently joins involving only relational and
arithmetic operators in join-clauses are pushed down to avoid any
potential join failure.  Also, only INNER and LEFT/RIGHT/FULL OUTER
joins are supported, and not SEMI and ANTI joins.

We observed that A JOIN (B JOIN C) doesn't work as expected on the
Hive server, so we convert it to the equivalent subquery form before
sending it to the remote server.

This patch also supports the partition-wise join push-down when the
table is partitioned.

FDW-176, Jeevan Ladhe, reviewed by Suraj Kharage and Jeevan Chalke.
Tested by Rajkumar Raghuwanshi.