Fix the fictitious test cases for non supported data type.
In the datatype.sql regression test file, there are cases to test
non-supported data types, assuming that those will throw an error for
respective column data types as they are not supported. But, since
each of the tests has an order by empno clause where empno is a non-
supported data type and is pushed as a first column, the next column
is never tested. Fix this by removing the order by clause.

Jeevan Ladhe, reviewed by Jeevan Chalke.