Push down ORDER BY to remote hdfs servers.
If possible, consider a path that adds the ORDER BY clause to the
remote SQL so that we get the ordered result set from the foreign
server itself.  It might help us to have an efficient merge join.

For some queries, we have observed that sorting on the Hive server
is very slow. Thus we have added a guc named enable_order_by_pushdown.
If it is set to true, then it pushes the order by operation to the
foreign server. Default is false.

FDW-179, Sravan Velagandula, reviewed by Suraj Kharage, tested by
Ajay Pal.