Check the shippability of sort clauses properly.
We were pushing the ORDER BY clauses to the remote side without
verifying whether the sort operator was safe to ship.  It resulted
in a wrong output when the sort operator isn't default for the sort
expression's type.  Fix the same.

FDW-563, Sravan Velagandula, reviewed by Suraj Kharage, Vaibhav Dalvi
and Jeevan Chalke.