Add enable_limit_pushdown guc to control LIMIT/OFFSET push-down at the session level.
LIMIT/OFFSET behavior is slightly different on HIVE and SPARK.  This
allows the user to control the push-down per the query/data pattern
and the client type.  On Apache Spark, support for the OFFSET clause
has been introduced in version 3.4.1.  So, for versions < 3.4.1, the
option either needs to be disabled or should not be used.  Default is
true.

FDW-645, Sravan Velagandula, reviewed by Suraj Kharage, tested by
Ajaykumar Pal.